requirements.txt
freetype-py
